@@ -1,7 +1,7 @@
 [package]
 build = "build.rs"
 description = "A framework for creating incredible standalone VR experiences"
-edition = "2018"
+edition = "2021"
 keywords = ["vr", "gamedev", "openxr", "vulkan"]
 license = "MIT OR Apache-2.0"
 name = "hotham"
@@ -14,19 +14,22 @@ anyhow = "1.0"
 ash = "0.33.2"
 bitflags = "1.3"
 cpal = "0.13.5"
-crossbeam = "0.8.1"
 ctrlc = {version = "3", features = ["termination"]}
 egui = "0.15"
 generational-arena = "0.2.8"
 glam = {features = ["mint", "serde", "approx"], version = "0.21.3"}
 gltf = {version = "1.0", features = ["KHR_lights_punctual", "KHR_materials_unlit", "names", "utils"], default-features = false}
+half = "2.1.0"
 hecs = "0.9.0"
+hotham-asset-client = {path = "../hotham-asset-client"}
 id-arena = "2.2.1"
 image = {version = "0.24.3", default-features = false, features = ["jpeg", "png"]}
 itertools = "0.10.0"
-ktx2 = "0.3"
+# ktx2 = "0.3"
+ktx2 = {git = "https://github.com/BVE-Reborn/ktx2"}
 memoffset = "0.6.5"
 mint = "0.5.6"
+notify-debouncer-mini = "0.2.1"
 oddio = "0.5"
 openxr = {features = ["loaded", "mint"], version = "0.17"}
 rapier3d = "0.14.0"
@@ -34,7 +37,7 @@ ruzstd = "0.3"
 serde = {version = "1.0", features = ["derive"]}
 symphonia = {version = "0.5", default-features = false, features = ["mp3"]}
 thiserror = "1.0"
-uuid = {version = "1.1", features = ["serde", "v4"]}
+tokio = {version = "1.0.1", default-features = false, features = ["rt"]}
 vk-shader-macros = "0.2.8"
 
 [target.'cfg(not(any(target_os = "macos", target_os = "ios")))'.dev-dependencies]
