@@ -319,18 +319,12 @@ fn get_collider_for_node(
     };
 
     // Build a collider using the mesh.
-    println!(
-        "[HOTHAM_ASSET_IMPORTER] Getting shape for {}",
-        collider_node_name
-    );
+    println!("[HOTHAM_ASSET_IMPORTER] Getting shape for {collider_node_name}");
     let shape = get_shape_from_mesh(mesh, import_context);
 
     // If this is a wall collider, ensure it's not a sensor.
     let collider = if collider_node_name.ends_with(WALL_COLLIDER_TAG) {
-        println!(
-            "[HOTHAM_ASSET_IMPORTER] Created wall collider for model {}",
-            collider_node_name
-        );
+        println!("[HOTHAM_ASSET_IMPORTER] Created wall collider for model {collider_node_name}");
         Collider {
             sensor: false,
             collision_groups: physics_context::WALL_COLLISION_GROUP,
@@ -339,10 +333,7 @@ fn get_collider_for_node(
             ..Default::default()
         }
     } else {
-        println!(
-            "[HOTHAM_ASSET_IMPORTER] Created sensor collider for model {}",
-            collider_node_name
-        );
+        println!("[HOTHAM_ASSET_IMPORTER] Created sensor collider for model {collider_node_name}");
         Collider {
             sensor: true,
             collision_groups: physics_context::SENSOR_COLLISION_GROUP,
@@ -361,7 +352,7 @@ fn find_wall_collider_for_node<'a>(
     import_context: &'a ImportContext,
 ) -> Option<(&'a str, gltf::Mesh<'a>)> {
     // Create a pattern to search for the collider's name, suffixed with the wall collider tag.
-    let wall_pattern = format!("{}{}", name, WALL_COLLIDER_TAG);
+    let wall_pattern = format!("{name}{WALL_COLLIDER_TAG}");
 
     // Iterate through each node to try and find the matching node, then fetch its mesh.
     import_context.document.nodes().find_map(|n| {
@@ -601,23 +592,18 @@ mod tests {
                 1,
                 23928,
                 [-0.06670809, 2.1408155, -0.46151406].into(),
-                Quat::from_xyzw(
-                    -0.09325116872787476,
-                    0.6883626580238342,
-                    0.006518156733363867,
-                    0.719318151473999,
-                ),
+                Quat::from_xyzw(-0.093_251_17, 0.688_362_66, 0.006_518_156_7, 0.719_318_15),
             ),
         ];
 
         for (name, id, indices_count, translation, rotation) in &test_data {
             let _ = models
                 .get(*name)
-                .expect(&format!("Unable to find model with name {}", name));
+                .unwrap_or_else(|| panic!("Unable to find model with name {name}"));
 
             let mut world = World::default();
-            let model = add_model_to_world(*name, &models, &mut world, None);
-            assert!(model.is_some(), "Model {} could not be added", name);
+            let model = add_model_to_world(name, &models, &mut world, None);
+            assert!(model.is_some(), "Model {name} could not be added");
 
             let model = model.unwrap();
             let (info, local_transform, mesh, ..) = world
@@ -650,7 +636,7 @@ mod tests {
             assert_relative_eq!(local_transform.translation, *translation,);
             assert_relative_eq!(local_transform.rotation, *rotation,);
             assert_eq!(&info.name, *name);
-            assert_eq!(&info.node_id, id, "Node {} has wrong ID", name);
+            assert_eq!(&info.node_id, id, "Node {name} has wrong ID");
         }
     }
 
