@@ -51,6 +51,7 @@ impl Texture {
         image_buf: &[u8],
         extent: &vk::Extent2D,
         array_layers: u32,
+        mip_count: u32,
         format: vk::Format,
         texture_usage: TextureUsage,
     ) -> Self {
@@ -62,7 +63,7 @@ impl Texture {
                 extent,
                 vk::ImageUsageFlags::SAMPLED | vk::ImageUsageFlags::TRANSFER_DST,
                 array_layers,
-                1,
+                mip_count,
                 component_mapping,
             )
             .unwrap();
@@ -72,7 +73,8 @@ impl Texture {
                 name,
                 vulkan_context,
                 image_buf,
-                1,
+                mip_count,
+                1, // note: use from_ktx2 if you want to import cubemaps
                 vec![image_buf.len() as u64 / array_layers as u64],
                 &image,
             )
@@ -145,7 +147,7 @@ impl Texture {
             )
             .unwrap();
         let index = render_context
-            .create_texture_image("Empty Texture", vulkan_context, &[], 1, vec![0], &image)
+            .create_texture_image("Empty Texture", vulkan_context, &[], 1, 1, vec![0], &image)
             .unwrap();
 
         Texture {
@@ -168,18 +170,44 @@ impl Texture {
         ktx2_data: &[u8],
         texture_usage: TextureUsage,
     ) -> Self {
+        println!("[HOTHAM_TEXTURE] Parsing KTX2 file {name}");
         let ktx2_image = parse_ktx2(ktx2_data);
+        println!(
+            "[HOTHAM_TEXTURE] KTX2 data: mip levels {}, array_layers: {}, faces: {}",
+            ktx2_image.mip_levels, ktx2_image.array_layers, ktx2_image.faces
+        );
 
-        Texture::new(
-            name,
-            vulkan_context,
-            render_context,
-            &ktx2_image.image_buf,
-            &ktx2_image.extent,
-            ktx2_image.array_layers.max(1) * ktx2_image.faces,
-            ktx2_image.format,
+        let component_mapping = get_component_mapping(&ktx2_image.format, &texture_usage);
+
+        // Right. Now we've got to do the array/mip count dance.
+        let image = vulkan_context
+            .create_image_with_component_mapping(
+                ktx2_image.format,
+                &ktx2_image.extent,
+                vk::ImageUsageFlags::SAMPLED | vk::ImageUsageFlags::TRANSFER_DST,
+                ktx2_image.faces,
+                ktx2_image.mip_levels,
+                component_mapping,
+            )
+            .unwrap();
+
+        let index = render_context
+            .create_texture_image(
+                name,
+                vulkan_context,
+                &ktx2_image.image_buf,
+                ktx2_image.mip_levels,
+                ktx2_image.faces,
+                ktx2_image.offsets,
+                &image,
+            )
+            .unwrap();
+
+        Texture {
+            image,
+            index,
             texture_usage,
-        )
+        }
     }
 
     /// Create a texture from an uncompressed image, like JPG or PNG.
@@ -223,25 +251,34 @@ impl Texture {
             &image.into_raw(),
             &extent,
             1,
+            1,
             format,
             texture_usage,
         )
     }
 }
 
-// Thin wrapper containing the information we need from a KTX2 file.
+/// Thin wrapper containing the information we need from a KTX2 file.
 #[derive(Debug, Clone)]
-pub(crate) struct KTX2Image {
+pub struct KTX2Image {
+    /// Image format
     pub format: vk::Format,
+    /// Image extents
     pub extent: vk::Extent2D,
+    /// Image buffer
     pub image_buf: Vec<u8>,
+    /// Offsets into the image buffer
     pub offsets: Vec<vk::DeviceSize>,
+    /// Number of mip levels
     pub mip_levels: u32,
+    /// Number of array layers
     pub array_layers: u32,
+    /// Number of faces
     pub faces: u32,
 }
 
-pub(crate) fn parse_ktx2(ktx2_data: &[u8]) -> KTX2Image {
+/// Parse some ktx2 data
+pub fn parse_ktx2(ktx2_data: &[u8]) -> KTX2Image {
     let ktx2_reader = ktx2::Reader::new(ktx2_data).unwrap();
     let header = ktx2_reader.header();
     let extent = vk::Extent2D {
@@ -252,26 +289,24 @@ pub(crate) fn parse_ktx2(ktx2_data: &[u8]) -> KTX2Image {
     let mut offsets = Vec::new();
 
     println!(
-        "[HOTHAM_TEXTURE] Importing KTX2 texture in {:?} format.",
-        header.format
+        "[HOTHAM_TEXTURE] Importing KTX2 texture in {:?} format with {} levels.",
+        header.format,
+        ktx2_reader.header().level_count,
     );
-    for level in ktx2_reader.levels() {
+    for mipmap_level in ktx2_reader.levels() {
         let len = match header.supercompression_scheme {
             // Lifted from Bevy, with Love:
             // https://github.com/bevyengine/bevy/blob/05e5008624b35f51cd6418acc745236be2cddd28/crates/bevy_render/src/texture/ktx2.rs#L62
             Some(ktx2::SupercompressionScheme::Zstandard) => {
-                let mut cursor = std::io::Cursor::new(level);
+                let mut cursor = std::io::Cursor::new(mipmap_level.data);
                 let mut decoder = ruzstd::StreamingDecoder::new(&mut cursor).unwrap();
                 decoder.read_to_end(&mut image_buf).unwrap()
             }
             None => {
-                image_buf.extend(level);
-                level.len()
+                image_buf.extend(mipmap_level.data);
+                mipmap_level.data.len()
             }
-            s => panic!(
-                "Unable to parse KTX2 file, unsupported supercompression scheme: {:?}",
-                s
-            ),
+            s => panic!("Unable to parse KTX2 file, unsupported supercompression scheme: {s:?}",),
         };
 
         let offset_increment = len as u32 / header.face_count;
@@ -318,10 +353,7 @@ fn get_format_from_mime_type(mime_type: &str) -> image::ImageFormat {
     match mime_type {
         "image/png" => image::ImageFormat::Png,
         "image/jpeg" => image::ImageFormat::Jpeg,
-        _ => panic!(
-            "Unable to import image - unsupported MIME type {}",
-            mime_type
-        ),
+        _ => panic!("Unable to import image - unsupported MIME type {mime_type}"),
     }
 }
 
