@@ -74,14 +74,14 @@ fn create_rigid_bodies(world: &mut hecs::World, physics_context: &mut PhysicsCon
     {
         if r.body_type == BodyType::Dynamic && parent.is_some() {
             panic!(
-                "[HOTHAM-PHYSICS] ERROR - Entities with parents cannot have dynamic rigid bodies: {:?}",
-                entity
+                "[HOTHAM-PHYSICS] ERROR - Entities with parents cannot have dynamic rigid bodies: {entity:?}"
             );
         }
 
         let mut rigid_body = RigidBodyBuilder::new(r.body_type.into())
             .additional_mass(r.mass)
             .position(global_transform.to_isometry())
+            .angvel(na_vector_from_glam(r.angular_velocity))
             .linvel(na_vector_from_glam(r.linear_velocity))
             .user_data(entity.to_bits().get() as _)
             .build();
@@ -163,8 +163,7 @@ fn update_rigid_bodies_from_world(physics_context: &mut PhysicsContext, world: &
 
         if body_type == BodyType::Dynamic && parent.is_some() {
             panic!(
-                "[HOTHAM-PHYSICS] ERROR - Entities with parents cannot have dynamic rigid bodies: {:?}",
-                entity
+                "[HOTHAM-PHYSICS] ERROR - Entities with parents cannot have dynamic rigid bodies: {entity:?}"
             );
         }
 
@@ -191,7 +190,7 @@ fn update_rigid_bodies_from_world(physics_context: &mut PhysicsContext, world: &
                 if world.get::<&Teleport>(entity).is_ok() {
                     command_buffer.remove_one::<Teleport>(entity);
                     let next_position = global_transform.to_isometry();
-                    println!("[HOTHAM_PHYSICS] Teleporting entity to {:?}", next_position);
+                    println!("[HOTHAM_PHYSICS] Teleporting entity to {next_position:?}");
                     rigid_body.set_position(next_position, true);
                 }
             }
@@ -205,17 +204,14 @@ fn update_rigid_bodies_from_world(physics_context: &mut PhysicsContext, world: &
                 if world.get::<&Teleport>(entity).is_ok() {
                     command_buffer.remove_one::<Teleport>(entity);
                     let next_position = global_transform.to_isometry();
-                    println!("[HOTHAM_PHYSICS] Teleporting entity to {:?}", next_position);
+                    println!("[HOTHAM_PHYSICS] Teleporting entity to {next_position:?}");
                     rigid_body.set_position(next_position, true);
                 }
 
                 // Apply one-shot components
                 if let Ok(additional_mass) = world.get::<&AdditionalMass>(entity).map(|a| a.value) {
                     command_buffer.remove_one::<AdditionalMass>(entity);
-                    println!(
-                        "[HOTHAM_PHYSICS] Applying additional mass of {:?}",
-                        additional_mass
-                    );
+                    println!("[HOTHAM_PHYSICS] Applying additional mass of {additional_mass:?}");
                     rigid_body.set_additional_mass(additional_mass, true);
                     rigid_body.recompute_mass_properties_from_colliders(&physics_context.colliders);
                 }
@@ -227,8 +223,7 @@ fn update_rigid_bodies_from_world(physics_context: &mut PhysicsContext, world: &
                         println!("[HOTHAM_PHYSICS] Attempted to apply impulse to rigid body with infinite mass. This is stupid and will do nothing.");
                     } else {
                         println!(
-                            "[HOTHAM_PHYSICS] Applying impulse of {:?} to rigid body with {} mass",
-                            impulse, mass
+                            "[HOTHAM_PHYSICS] Applying impulse of {impulse:?} to rigid body with {mass}"
                         );
                         rigid_body.apply_impulse(na_vector_from_glam(impulse), true);
                     }
